Split audio out of Clyde
